requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorzero==2.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gpiozero==1.6.2
httplib2==0.22.0
idna==3.4
lgpio==0.0.0.2
oauth2client==4.1.3
oauthlib==3.2.2
picamera==1.13
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyDrive==1.3.1
pyparsing==3.0.9
pypng==0.20220715.0
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
RPIO==0.10.0
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
